extract into filters file
